。

+++ server/src/main/java/org/luncert/facedetect/repo/TeacherRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TeacherRepo extends CrudRepository<Teacher, Integer> {
+public interface TeacherRepo extends CrudRepository<Teacher, Long> {
 }
